drivers: nvcsi: Fix power sequence race condition
Fix race condition in power sequence of nvcsi, where
the power on returns zero but waits to perform
calibration sequence and depends on RTCPU to be powered
on. If RTCPU is not powered on before power off request
is issued then power off sequence will execute before
completion of power on. The atomic variable context
must be set after the power on/off sequence is complete.

Bug 200427615

Change-Id: I336f91415836143285d897ee81c3741ca6291f25
Signed-off-by: Bhanu Murthy V <bmurthyv@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1772893
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: David Bang <dbang@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Frank Chen <frankc@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>