drivers: camera: protect channel buffer count
Operations that modify a channel's buffer count
(freeing and adding of ring buffers) are not exclusive and can cause
race conditions. This patch places these operations in critical
sections to grant exclusive modification.

This patch also modifies the logic of freeing ring buffers to only
execute when the number of buffers to free is greater than zero.
This prevents infinite freeing loops when the number of buffers to
free becomes corrupted.

Bug 200409712
Bug 200422550

Change-Id: I7646a49bf6a14c4a4f8a3597e72d4bccdc8a0336
Signed-off-by: Ian Kaszubski <ikaszubski@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1774549
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Bhanu Murthy V <bmurthyv@nvidia.com>
Reviewed-by: Bhushan Rayrikar <brayrikar@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Frank Chen <frankc@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>