cpufreq: clamp boot frequency in core cpufreq
Set the flag to let cpufreq check the boot frequency against
available frequencies in frequency table. clamp to a valid
entry if the boot frequency is not in the table.

Remove the relevant code in tegra cpufreq driver for t186 and
t194. This is no longer needed.

bug 2255138

Change-Id: I47949ef4c641d9c2406480e1efafd71821632d3f
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1775429
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>