video: tegra: host: unify the VI clock name
Sync up the VI clock name as "vi" among different platforms. In order to
do this on Tegra210, we also need to fix the DT clock ID to the clock
that we want to use as "vi" clock. For Tegra210, which is
"vi_v4l2.cbus".

Bug 200418280

Change-Id: Iab278bc7792848a1c2b32495a485119a2c925078
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1776742
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Bhanu Murthy V <bmurthyv@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>