t194: dt: Add all FB regions to the display heads
Add each of the FB direct mapped regions to the
display heads. This is necessary because the
display heads share SIDs, so when one device probes
all of the heads will be using the IOMMU.
This means one head probes we need all of the FBs
to be mapped.

Bug 2116047

Change-Id: I4f2df0615f9575896d5ede5914d0748f5b4274f3
Signed-off-by: Ajay Nandakumar <anandakumarm@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1777733
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>