nvdisp: adjust worst-case DVFS latencies
For IMP, the DVFS switch latency for each EMC frequency needs
to be the max of the latency at that particular frequency
*and* all of the frequencies above it. Although IMP enforces an EMC
floor value (max of the minimum EMC needed to pass LA/PTSA vs. the
required total ISO bandwidth), MC is still free to switch the EMC
frequency to any higher point. For Xavier, the periodic training that's
done for some of the upper EMC frequencies can cause the worst-case
dead time to spike up to 80us, which is why it's especially important
to get the DVFS buffering right.

This patch updates tegra_nvdisp_init_common_imp_data() so that
it correctly pre-processes the latency values in the EMC DVFS table
that the driver queries from bpmp.

Bug 200412594
TDS-3621

Change-Id: Id4b96cb7602e86714c221ff3f84ef257cc08cd7c
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1781200
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>