thermal: fix get_trend
The thermal framework programs a trip point only if the current temp
is _greater_ than the trip point. There could therefore be a corner
case where the temperature readback after crossing the trip is equal
to the trip temperature. In such a condition the sensor driver must
notify the thermal governor (via the get_trend) to start polling the
temperature so that the trip can be reprogrammed once the temperature
goes below the trip point.

Bug 200430120

Change-Id: I9d76084d036d7933ce55d5174b5da06f925d9484
Signed-off-by: Srikar Srimath Tirumala <srikars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1780341
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Rajkumar Kasirajan <rkasirajan@nvidia.com>
Tested-by: Rajkumar Kasirajan <rkasirajan@nvidia.com>
Reviewed-by: Winnie Hsu <whsu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>