video: fbconsole: API to check if fbcon is active
Added an API which helps determine if fbconsole is currenlty
active on the foreground console. This is useful when user
needs to figure out if fbconsole or X is the active client.

Also, fb window update during unblanking is needed only if
fbconsole is active. Hence, added an extra check to set
dc->blanked to true only if fbconsole is currently active
during powerdown.

bug 200412594

Change-Id: I0f1596666aa640103fe229d131ac9c5f97b62af8
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1783397
Reviewed-by: Shu Zhong <shuz@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: Tow Wang <toww@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>