dts: t210: Add iommu-map and iommu-map-mask
Add iommu-map and iommu-map-mask properties to allow PCI devices to
register with of-iommu.

Based on the map and map-mask properties, a PCI RID is converted to
swgroup ID and subsequently registered with the iomm_fwspec via the
of_xlate call.

Since, on T210, the PCIe devices use 'AFI' swgroup, which happens to be
'0' in value, adding a 0x0 map-mask works to first mask the RID to '0',
and then, add AFI swgroup value (specified in map property) to deduce
the swgroup for PCI devices. Later, call to of_xlate with the same value
lets the smmu driver register swgroup for PCI with the iommu_fwspec.

Bug 200419422
Bug 200420606

Change-Id: Ia5641b28ad96fe851e0dfa76c536fe845d0c52d1
Signed-off-by: Navneet Kumar <navneetk@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1782072
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>