video: tegra: fix pointer handling during hotplug
KASAN complained of multiple "use after free" and "freeing invalid
pointer" errors during HDMI/DP hotplug. Turned out that we were
freeing modedb pointer during both plug and unplug (and setting it
to NULL) which caused the inconsistent behaviour.
To fix it, moved the modedb free instructions under unplug path so
that we avoid freeing it again when a sink is plugged.

bug 200433449

Change-Id: Ic0e7901819e55542c2ec62e6380990c15f78b3bc
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1785293
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>