scsi: ufs: remove cd_irq override
Interrupt number is allocated with gpio_to_irq number and saved in
cd_irq which is used during suspend/resume for enabling/disabling irq.
Do not override cd_irq with the error value.

Also, disabled suspend-resume for the condition when the card is not
present.

Bug 200433759
Bug 200377955

Change-Id: Icff234c82b2b4b83a21734b91b13e433ff3b262a
Signed-off-by: Sri Sai Ram Tangirala <stangirala@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1784670
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>