video: tegra: nvmap: Remove dependencies on nvmap_priv.h
Remove each of the file's dependencies on nvmap_priv.h.
This requires creating more APIs so the direct structures
do not need to be referenced.

All of the C structures have been pushed into .c files
except for nvmap_dev which is referenced in too many
places. This will be removed in a later commit.

TMM-123

Change-Id: I480f630653da65a1e31be6c907bc4ae5049e2e1e
Signed-off-by: David Gilhooley <dgilhooley@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1810134
GVS: Gerrit_Virtual_Submit
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>