Revert "arm64: Increase the max granular size"
This reverts commit 97303480753e48fb313dc0e15daaf11b0451cdb8.

The change raised the min kmalloc size to 128 Bytes which creates
internal fragmentations and memory waste. And there seems to be a
performance regression on the network packets allocation, reported
in the upstream: https://patchwork.kernel.org/patch/10249341/

According to the original change, the change was made for Cavium's
ThunderX platforms, so it should be safe for us to revert.

After reverting back to 6, the total memory usage via kmalloc()
reduced from 21.9MB to 18.9MB, when testing Netflix playback.

Bug 2284925

Change-Id: Ic1b862ecad9c636e0741f60f8d73dc44dec82451
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1803456
GVS: Gerrit_Virtual_Submit
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>