sched: cpufreq: Reduce overflow possibility
During the calculation for next frequency, there is a possiblity that
the intermediate value obtained after multiplying current freq by
capacity_margin will overflow UINT_MAX. This change-set eliminates that
possibility by casting freq to a temporary 64-bit variable and doing the
calculation in 64-bit. The final value is then cast to 32-bit because it
is guaranteed to be within UINT_MAX range as long as freq is less than
UINT_MAX / 2 which it always is.

Bug 2106745

Change-Id: If83d6c7104e813bb70bfdf657886d5ca458c35a1
Signed-off-by: Waqar Ali <waqara@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1794478
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bo Yan <byan@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>