arm64: configs: make AHCI as part of the kernel
When a SATA Port Mulitplier is connected along with a SATA drive,
the Marvell SATA Controller is sending an interrupt to PCIE.
The pcie driver will not handle this interrupt and
it is expected that SATA AHCI handle it.
But on L4T currently AHCI driver is built as kernel module and
is not loaded until rootfs is mounted. Since PCIE driver is not
handling this interrupt  there is continuous generation of interrupts
which is stalling the CPU.

Bug 200417967

Change-Id: I59d6afdae516696a7138255db63d8aa2db161ca6
Signed-off-by: Preetha Chandru R <pchandru@nvidia.com>
Signed-off-by: Vidya Sagar <vidyas@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1801755
(cherry-picked from commit a134d753e9627f6eafb618ca9f0c0fcc238a7bac)
Reviewed-on: https://git-master.nvidia.com/r/1805977
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>