arm: smmu: tegra: Fix ptdump address values
Fix smmu_ptdump_show() so that the body of each loop always increments
addr by the appropriate amount, irrespective of whether any child tables
or PTEs are present or not. This removes the need for bizarre special
cases, and corrects the output in the debugfs file. For example, in the
existing code, if 4M of entries at IOVA 0xff800000 were valid, the debugfs
code would print 2M of entries at IOVA ff800000 and 2M of entries at
ffc00000!

Bug 2312071

Change-Id: I6d0df7d29909782bfbd950fea49c9d7a23ef0ccc
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1806368
Reviewed-by: Puneet Saxena <puneets@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: David Gilhooley <dgilhooley@nvidia.com>
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>