cpufreq: eliminate Sparse errors
If the explicit restriction for endianness needs to be applied
on the MRQ data type, the type needs to be fixed in relevant header
for BPMP ABI. But this is not necessary since kernel and BPMP SW always
run in little endian. Therefore, remove the cpu_to_le32 conversion to
eliminate Sparse errors.

Bug 200434802

Change-Id: I1b753af972c070662d151279823a52a61d6e452b
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1807578
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>