firmware: tegra: explicilty cast little-endians to uint32_t
The BPMP-FW firmware header uses little endian byte order in the
command structs. However, the struct members are declared as uint32_t
and sparse emits warnings as cpu_to_le32() return type is __le32.

To avoid these warnings, cast the return value of cpu_to_le32()
explicitly to uint32_t. It is arguable that more correct way to
resolve the warnings would be to change the ABI header data
types. However, this is unfavorable as the same header is shared
across multiple operating systems and __le32 is not standard data
type.

Bug 200434802

Change-Id: I3b27b0086a7b6aeb9707e72c21e08bbf89fe8a7b
Signed-off-by: Timo Alho <talho@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1811967
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>